backend/requirements.txt
Django==2.2.16
djangorestframework==3.12.4
djoser==2.1.0
django-filter==2.4.0
pillow==9.5.0
sorl-thumbnail==12.9.0
drf-extra-fields==3.4.1
reportlab==3.6.8
django-cors-headers==3.11.0
psycopg2-binary
gunicorn==20.1.0
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
python-dotenv==0.19.2




